+++ lib/app/cubit/sign_in/sign_in_state.dart
@@ -1,0 +1,12 @@
+import 'package:equatable/equatable.dart';
+
+abstract class SignInState extends Equatable {
+  const SignInState();
+
+  @override
+  List<Object?> get props => [];
+}
+
+class SignInPrimaryState extends SignInState {
+  const SignInPrimaryState() : super();
+}
