+++ lib/app/definition/string.dart
@@ -1,7 +1,8 @@
 abstract class DigPageName {
   static const String root = '/';
   static const String splash = '/splash';
-  static const String login = '/login';
+  static const String signIn = '/sign_in';
   static const String home = '/home';
   static const String backUpYourWallet = '/back_up_your_wallet';
-}
+  static const String createAccount = '/create_account';
+  static const String importAccount = '/import_account';}
