+++ lib/app/route/dig_route.dart
@@ -1,7 +1,7 @@
 import 'package:dig_mobile_app/app/definition/string.dart';
 import 'package:dig_mobile_app/app/page/create_wallet_flow/back_your_wallet/backup_your_wallet_page.dart';
 import 'package:dig_mobile_app/app/page/home/home_page.dart';
-import 'package:dig_mobile_app/app/page/login/login_page.dart';
+import 'package:dig_mobile_app/app/page/sign_in/sign_in_page.dart';
 import 'package:dig_mobile_app/app/page/splash/splash.dart';
 import 'package:flutter/material.dart';
 
@@ -16,7 +16,7 @@ class DigRoute {
       case DigPageName.root:
       case DigPageName.splash:
         return _materialPage(const SplashPage(), settings);
-      case DigPageName.login:
+      case DigPageName.signIn:
         return _materialPage(const LoginPage(), settings);
       case DigPageName.home:
         return _materialPage(const HomePage(), settings);
