+++ lib/di/di.config.dart
@@ -8,7 +8,7 @@ import 'package:get_it/get_it.dart' as _i1;
 import 'package:injectable/injectable.dart' as _i2;
 
 import '../app/cubit/active_account/active_account_cubit.dart' as _i3;
-import '../app/cubit/login/login_cubit.dart' as _i4;
+import '../app/cubit/sign_in/sign_in_cubit.dart' as _i4;
 import '../app/cubit/splash/splash_cubit.dart'
     as _i5; // ignore_for_file: unnecessary_lambdas
 
@@ -18,7 +18,7 @@ _i1.GetIt $initGetIt(_i1.GetIt get,
     {String? environment, _i2.EnvironmentFilter? environmentFilter}) {
   final gh = _i2.GetItHelper(get, environment, environmentFilter);
   gh.factory<_i3.ActiveAccountCubit>(() => _i3.ActiveAccountCubit());
-  gh.factory<_i4.LoginCubit>(() => _i4.LoginCubit());
+  gh.factory<_i4.SignInCubit>(() => _i4.SignInCubit());
   gh.factory<_i5.SplashCubit>(() => _i5.SplashCubit());
   return get;
 }
